refactor: clear and optimize code
