+++ src/components/admin/client/datagrid/index.js
@@ -8,7 +8,7 @@ import { BsToggleOff, BsToggleOn } from 'react-icons/bs'
 import { setStatusClient } from '../../../../store/client/client.action'
 import { listByIdUserAssetAction } from '../../../../store/transaction/transaction.action'
 import Loading from '../../../loading/index'
-import { BoxTable } from './DatagridElements'
+import { BoxTable } from '../../../datagrid/styled'
 import ListFinancialAssets from '../financial_assets'
 import ListFinancialTransaction from '../financial_transaction'
 import ListClientDetails from '../clients_details'
