+++ src/components/admin/deposit/datagrid/index.js
@@ -4,7 +4,7 @@ import { DataGrid } from '@material-ui/data-grid'
 import { BsFillCartFill } from 'react-icons/bs'
 import { More as MoreIcon } from '@mui/icons-material'
 import { IconButton, Tooltip } from '@material-ui/core'
-import { BoxTable } from './DatagridElements'
+import { BoxTable } from '../../../datagrid/styled'
 import Loading from '../../../loading/index'
 import Form from '../deposit'
 import { listByIdUserDepositAction } from '../../../../store/transaction/transaction.action'
