+++ src/components/admin/transaction/datagrid/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { DataGrid } from '@material-ui/data-grid'
-import { BoxTable } from './DatagridElements'
+import { BoxTable } from '../../../datagrid/styled'
 import Loading from '../../../loading/index'
 
 const DataList = ({ data, modal, loading }) => {
