+++ src/components/datagrid/index.js
@@ -1,10 +1,11 @@
 import React from 'react'
 import { DataGrid } from '@material-ui/data-grid'
-import styled from 'styled-components'
+import { BoxTable } from './styled'
+import Loading from '../loading/index'
 
 const DataList = ({ data, columns, loading }) => {
   if (loading) {
-    return <p>carregando...</p>
+    return <Loading />
   }
 
   return (
@@ -15,8 +16,3 @@ const DataList = ({ data, columns, loading }) => {
 }
 
 export default DataList
-
-const BoxTable = styled.div`
-  height: 600px;
-  width: 100%;
-`
