+++ src/components/loading/styled.js
