+++ src/components/title/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { Typography } from '@material-ui/core'
-import { Title } from './TitleElements'
+import { Title } from './styled'
 
 const Head = ({ title, actions: Actions }) => {
   return (
