+++ src/components/title/styled.js
